+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+- `SeekBarConfig.markerUpdateIntervalMs` to customize the update interval of the TimeLine markers in a live stream
+
+### Fixed
+- `TimelineMarker`s move unexpectedly during time-shifting in a live stream
+
 ## [3.104.0] - 2025-12-09
 
 ### Fixed